refactor(app): :recycle: separate main class
